+++ src/main/java/org/realityforge/replicant/client/json/gwt/GwtDataLoaderService.java
@@ -3,7 +3,7 @@ package org.realityforge.replicant.client.json.gwt;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.core.client.Scheduler;
 import javax.annotation.Nonnull;
-import org.realityforge.replicant.client.AbstractDataLoaderService;
+import org.realityforge.replicant.client.transport.AbstractDataLoaderService;
 import org.realityforge.replicant.client.ChangeSet;
 import org.realityforge.replicant.client.transport.ClientSession;